Refactors and code improvements
